Initial work on config file, kapal up, move cmds into own package
