+++ kapal.go
@@ -3,86 +3,11 @@ package main
 import (
 	"fmt"
 	"os"
-	"os/exec"
-	"path"
-	"bytes"
-	"strings"
 	"github.com/codegangsta/cli"
-	"github.com/hoodiez/kapal/btrfs"
 	"github.com/hoodiez/kapal/sysfs"
+	"github.com/hoodiez/kapal/cmds"
+	"github.com/hoodiez/kapal/config"
 )
-
-func CreateVolume (pool string, name string, dockerize bool, dockername string, dockervol string) {
-	err := btrfscmd.SubvolumeCreate(pool,name)
-	if err != nil {
-		fmt.Println("Failed creating Volume in the file system")
-	} else {
-		/*Todo: should be using Docker API and or it's own wrapper Lib*/
-		if dockerize {
-			var cmd *exec.Cmd
-			var containervol string
-			var out bytes.Buffer
-			if dockervol != "" {
-				containervol = dockervol
-				if strings.HasPrefix(containervol, "/") == false {
-					containervol = "/"+containervol
-				}
-			} else {
-				containervol = "/data"
-			}
-			if dockername != "" {
-				cmd = exec.Command("docker","create","-v",path.Join(pool,name)+":"+containervol,"--name",dockername,"ubuntu")
-			} else {
-				cmd = exec.Command("docker","create","-v",path.Join(pool,name)+":"+containervol,"ubuntu")
-			}
-			cmd.Stdout = &out
-			err2 := cmd.Run()
-
-			if err2 != nil {
-				fmt.Println("Error creating Docker Data Volume Container")
-			} else {
-				fmt.Print(out.String())
-			}
-		}
-	}
-}
-
-func CloneVolume (pool string, source string, target string, readonly bool, dockerize bool, dockername string, dockervol string) {
-	err := btrfscmd.SubvolumeSnapshot(pool,source,target,readonly)
-	if err != nil {
-                fmt.Println("Failed cloning Volume in the file system")
-        } else {
-                /*Todo: should be using Docker API and or it's own wrapper Lib*/
-                if dockerize {
-                        var cmd *exec.Cmd
-                        var containervol string
-			var out bytes.Buffer
-			
-                        if dockervol != "" {
-                                containervol = dockervol
-                                if strings.HasPrefix(containervol, "/") == false {
-                                        containervol = "/"+containervol
-                                }
-                        } else {
-                                containervol = "/data"
-                        }
-                        if dockername != "" {
-                                cmd = exec.Command("docker","create","-v",path.Join(pool,target)+":"+containervol,"--name",dockername,"ubuntu")
-                        } else {
-                                cmd = exec.Command("docker","create","-v",path.Join(pool,target)+":"+containervol,"ubuntu")
-                        }
-			cmd.Stdout = &out
-                        err2 := cmd.Run()
-
-                        if err2 != nil {
-                                fmt.Println("Error creating Docker Data Volume Container")
-                        } else {
-				fmt.Print(out.String())
-			}
-
-                }
-        }
-}
 
 func main() {
 	version := "0.0.1"
@@ -92,6 +17,13 @@ func main() {
 	app.Version = version 
 	app.Authors = []cli.Author{cli.Author{Name: "hoodiez", Email:"https://github.com/hoodiez"},cli.Author{Name: "jteso",Email: "https://github.com/jteso"}}
 	app.Commands = []cli.Command {
+		{
+			Name: "up",
+			Usage: "Read Kapalfile and intialize device, pool and volumes configuration",
+			Action: func(c *cli.Context) {
+				config.ReadKapalFile(c.Args().First())
+			},
+		},
 		{
 			Name: "device",
 			Aliases: []string{"d"},
@@ -146,7 +78,7 @@ func main() {
                                 Usage: "Create a new volume in a pool",
                                 Action: func(c *cli.Context) {
                                         fmt.Println("Create Volume: ", c.String("pool"), c.String("name"))
-					CreateVolume(c.String("pool"),c.String("name"),c.Bool("dockerize"),c.String("dockername"),c.String("dockervol"))
+					cmds.CreateVolume(c.String("pool"),c.String("name"),c.Bool("dockerize"),c.String("dockername"),c.String("dockervol"))
                                 },
                                 Flags: []cli.Flag {
                                  cli.StringFlag{
@@ -206,7 +138,7 @@ func main() {
 				Usage: "Clone a volume into a another volume in a pool",
 				Action: func(c *cli.Context) {
                                         fmt.Println("Clone Volume: ", c.String("pool"),c.String("source"),c.String("target"))
-					CloneVolume(c.String("pool"),c.String("source"),c.String("target"),c.Bool("readonly"),c.Bool("dockerize"),c.String("dockername"),c.String("dockervol"))
+					cmds.CloneVolume(c.String("pool"),c.String("source"),c.String("target"),c.Bool("readonly"),c.Bool("dockerize"),c.String("dockername"),c.String("dockervol"))
                                 },
                                 Flags: []cli.Flag {
                                  cli.StringFlag{
